pyproject.toml
[tool.poetry]
name = "fabric-mcp-demo"
version = "1.0.0"
description = "Enterprise Multi-Tenant MCP Server with FoundryIQ Integration"
authors = ["Your Organization <support@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
azure-identity = "^1.15.0"
azure-keyvault-secrets = "^4.7.0"
azure-storage-blob = "^12.19.0"
azure-monitor-opentelemetry = "^1.2.0"
azure-mgmt-costmanagement = "^4.0.1"
azure-communication-email = "^1.0.0"
azure-communication-sms = "^1.0.1"
redis = "^5.0.1"
fastapi-limiter = "^0.1.6"
httpx = "^0.26.0"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
msal = "^1.26.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dateutil = "^2.8.2"
pytz = "^2023.3"
pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
